requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# MLOps — experiment tracking & model registry
mlflow>=2.10.0

# Data download
kagglehub>=0.2.0

# Model serving (FastAPI)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Drift monitoring
evidently>=0.4.0

# Data versioning (Azure remote)
dvc[azure]>=3.0.0
